INT-2031 changed the priority type back to Integer

+++ spring-integration-core/src/main/java/org/springframework/integration/MessageHeaders.java
@@ -136,8 +136,8 @@ public final class MessageHeaders implements Map<String, Object>, Serializable {
 		return (sequenceSize != null ? sequenceSize : 0);
 	}
 
-	public Long getPriority() {
-		return this.get(PRIORITY, Long.class);
+	public Integer getPriority() {
+		return this.get(PRIORITY, Integer.class);
 	}
 
 	@SuppressWarnings("unchecked")
